refactor: fix circular dependency
