+++ wormhole-connect/src/views/Bridge/Inputs.tsx/To.tsx
@@ -1,113 +1,27 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { BigNumber } from 'ethers';
-import { makeStyles } from 'tss-react/mui';
 import { RootState } from '../../../store';
-import { CircularProgress, Link, Typography } from '@mui/material';
 import { setToNetworksModal } from '../../../store/router';
-import { TransferWallet, signSolanaTransaction } from '../../../utils/wallet';
-import { getWrappedToken, getWrappedTokenId } from '../../../utils';
-import { joinClass } from '../../../utils/style';
-import { ATTEST_URL, TOKENS } from '../../../config';
-import { getBalance, getForeignAsset, solanaContext } from '../../../sdk';
-import {
-  formatBalance,
-  setAssociatedTokenAddress,
-  setForeignAsset,
-} from '../../../store/transfer';
+import { TransferWallet } from '../../../utils/wallet';
+import { getWrappedToken } from '../../../utils';
+import { TOKENS } from '../../../config';
+import { getBalance } from '../../../sdk';
+import { formatBalance } from '../../../store/transfer';
 
 import Inputs from './Inputs';
 import Input from './Input';
 import Select from './Select';
 import InputTransparent from '../../../components/InputTransparent';
-
-const useStyles = makeStyles()((theme) => ({
-  associatedTokenWarning: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'start',
-  },
-  link: {
-    textDecoration: 'underline',
-    opacity: '0.8',
-    padding: '4px 0',
-    cursor: 'pointer',
-    '&:hover': {
-      opacity: '1',
-    },
-  },
-  disabled: {
-    cursor: 'not-allowed',
-    opacity: '0.6',
-    '&:hover': {
-      opacity: '0.6',
-    },
-  },
-  inProgress: {
-    marginRight: '8px',
-  },
-  error: {
-    color: theme.palette.error[500],
-    marginTop: '4px',
-  },
-}));
-
-type Props = {
-  createAssociatedTokenAccount: any;
-};
-function AssociatedTokenWarning(props: Props) {
-  const { classes } = useStyles();
-  const [inProgress, setInProgress] = useState(false);
-  const [error, setError] = useState('');
-
-  const createAccount = async () => {
-    // if `createAccount` is already in progress, disable function
-    if (inProgress) return;
-    setInProgress(true);
-    setError('');
-    try {
-      await props.createAssociatedTokenAccount();
-      setError('');
-    } catch (e) {
-      setError('Encountered an error, please try again.');
-      console.error(e);
-    } finally {
-      setInProgress(false);
-    }
-  };
-
-  return (
-    <div className={classes.associatedTokenWarning}>
-      No associated token account exists for your wallet on Solana. You must
-      create it before proceeding.
-      {error && <div className={classes.error}>{error}</div>}
-      <div
-        className={joinClass([classes.link, inProgress && classes.disabled])}
-        onClick={createAccount}
-      >
-        {inProgress && (
-          <CircularProgress size={18} className={classes.inProgress} />
-        )}
-        Create account
-      </div>
-    </div>
-  );
-}
+import TokenWarnings from './TokenWarnings';
 
 function ToInputs() {
   const dispatch = useDispatch();
   const [balance, setBalance] = useState(undefined as string | undefined);
-  const [warnings, setWarnings] = useState([] as any[]);
 
-  const {
-    validations,
-    fromNetwork,
-    toNetwork,
-    token,
-    amount,
-    foreignAsset,
-    associatedTokenAddress,
-  } = useSelector((state: RootState) => state.transfer);
+  const { validations, fromNetwork, toNetwork, token, amount } = useSelector(
+    (state: RootState) => state.transfer,
+  );
   const { sending, receiving } = useSelector(
     (state: RootState) => state.wallet,
   );
@@ -132,29 +46,6 @@ function ToInputs() {
     );
   }, [tokenConfig, fromNetwork, toNetwork, receiving.address]);
 
-  // check if the destination token contract is deployed
-  useEffect(() => {
-    const checkWrappedTokenExists = async () => {
-      if (!toNetwork || !token) {
-        dispatch(setForeignAsset(''));
-        return;
-      }
-
-      const tokenConfig = TOKENS[token];
-      const tokenId = tokenConfig.tokenId
-        ? tokenConfig.tokenId
-        : getWrappedToken(tokenConfig).tokenId;
-
-      if (!tokenId) {
-        throw new Error('Could not retrieve target token info');
-      }
-
-      const address = await getForeignAsset(tokenId, toNetwork);
-      dispatch(setForeignAsset(address || ''));
-    };
-    checkWrappedTokenExists();
-  }, [toNetwork, token, dispatch]);
-
   // token display jsx
   const symbol = tokenConfig && getWrappedToken(tokenConfig).symbol;
   const selectedToken = tokenConfig
@@ -175,107 +66,6 @@ function ToInputs() {
     </Input>
   );
 
-  const createAssociatedTokenAccount = useCallback(async () => {
-    if (!receiving.address || !token)
-      throw new Error(
-        'Must fill in all fields before you can create a token account',
-      );
-    if (!foreignAsset)
-      throw new Error(
-        'The token must be registered on Solana before an associated token account can be created',
-      );
-    const tokenId = getWrappedTokenId(tokenConfig);
-    const tx = await solanaContext().createAssociatedTokenAccount(
-      tokenId,
-      receiving.address,
-      'finalized',
-    );
-    // if `tx` is null it means the account already exists
-    if (!tx) return setWarnings([]);
-    await signSolanaTransaction(tx, TransferWallet.RECEIVING);
-
-    let accountExists = false;
-    let retries = 0;
-    return await new Promise((resolve) => {
-      const checkAccount = setInterval(async () => {
-        if (accountExists || retries > 20) {
-          clearInterval(checkAccount);
-          resolve(true);
-        } else {
-          accountExists = await checkSolanaAssociatedTokenAccount();
-          retries += 1;
-        }
-      }, 1000);
-    });
-  }, [token, receiving, foreignAsset, tokenConfig]);
-
-  // destination token warnings
-  const tokenWarning = useMemo(
-    () => (
-      <Typography>
-        This token is not registered, you must{' '}
-        <Link target={'_blank'} variant="inherit" href={ATTEST_URL}>
-          register
-        </Link>{' '}
-        it before you continue. Newly registered tokens will not have liquid
-        markets.
-      </Typography>
-    ),
-    [],
-  );
-  const associatedTokenWarning = useMemo(
-    () => (
-      <AssociatedTokenWarning
-        createAssociatedTokenAccount={createAssociatedTokenAccount}
-      />
-    ),
-    [createAssociatedTokenAccount],
-  );
-
-  // the associated token account address is deterministic, so we still
-  // need to check if there is an account created for that address
-  const checkSolanaAssociatedTokenAccount =
-    useCallback(async (): Promise<boolean> => {
-      if (!foreignAsset) return false;
-      let tokenId = tokenConfig.tokenId || getWrappedTokenId(tokenConfig);
-      const account = await solanaContext().getAssociatedTokenAccount(
-        tokenId,
-        receiving.address,
-      );
-      if (account) {
-        dispatch(setAssociatedTokenAddress(account.toString()));
-        setWarnings([]);
-        return true;
-      } else {
-        setWarnings([associatedTokenWarning]);
-        return false;
-      }
-    }, [
-      foreignAsset,
-      tokenConfig,
-      receiving,
-      dispatch,
-      associatedTokenWarning,
-    ]);
-
-  useEffect(() => {
-    if (!toNetwork || !token || !receiving.address) return setWarnings([]);
-    if (!foreignAsset) return setWarnings([tokenWarning]);
-    if (toNetwork === 'solana') {
-      checkSolanaAssociatedTokenAccount();
-    } else {
-      setWarnings([]);
-    }
-  }, [
-    toNetwork,
-    token,
-    foreignAsset,
-    receiving,
-    associatedTokenAddress,
-    tokenWarning,
-    checkSolanaAssociatedTokenAccount,
-  ]);
-
   return (
     <Inputs
       title="To"
@@ -289,7 +79,7 @@ function ToInputs() {
       tokenInput={tokenInput}
       amountInput={amountInput}
       balance={balance}
-      warnings={warnings}
+      warning={<TokenWarnings />}
     />
   );
 }
